fix(st-spi): avoid overflow in stm32_ospi_dtr_calibrate()
If bus frequency is lower than 15.25MHz in stm32_ospi_dtr_calibrate, the
computation of period_ps will overflow an uint16_t, and will then change
the its value. To avoid this, use uint32_t for period_ps, and change
stm32mp_syscfg_dlyb_init() and stm32mp_syscfg_find_byp_cmd() prototypes
to use uint32_t for the period_ps parameter.

This was found when enabling -Wconversion warning:
drivers/st/spi/stm32_ospi.c: In function 'stm32_ospi_dtr_calibrate':
drivers/st/spi/stm32_ospi.c:129:33: warning: conversion from 'unsigned int' to 'uint16_t'
  {aka 'short unsigned int'} may change value [-Wconversion]
  129 | #define _OSPI_NSEC_PER_SEC      1000000000U
      |                                 ^~~~~~~~~~~
drivers/st/spi/stm32_ospi.c:352:29: note: in expansion of macro '_OSPI_NSEC_PER_SEC'
  352 |                 period_ps = _OSPI_NSEC_PER_SEC / (bus_freq / 1000U);
      |                             ^~~~~~~~~~~~~~~~~~

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/557400
Domain-Review: Christophe KERELLO <christophe.kerello@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Tested-by: Yann GAUTIER <yann.gautier@foss.st.com>

+++ drivers/st/spi/stm32_ospi.c
@@ -336,7 +336,7 @@ static int stm32_ospi_mm(uint8_t *buf, uint32_t nbytes, size_t addr)
 static int stm32_ospi_dtr_calibrate(uint32_t prescaler, unsigned int bus_freq,
 				    int (*check_transfer)(void))
 {
-	uint16_t period_ps = 0U;
+	uint32_t period_ps = 0U;
 	uint8_t window_len = 0U;
 	int ret;
 	bool bypass_mode = false;
