+++ plat/st/stm32mp2/include/stm32mp2_private.h
@@ -14,7 +14,7 @@ void stm32mp_syscfg_mask_potential_tamper_enable(void);
 void stm32mp_syscfg_mask_potential_tamper_disable(void);
 size_t stm32mp_syscfg_get_mm_size(uint8_t bank);
 int stm32mp_syscfg_dlyb_init(uint8_t bank, bool bypass_mode,
-			      uint16_t period_ps);
+			     uint32_t period_ps);
 int stm32mp_syscfg_dlyb_find_tap(uint8_t bank, int (*check_transfer)(void),
 				  bool rx_only, uint8_t *window_len);
 int stm32mp_syscfg_dlyb_set_cr(uint8_t bank, uint32_t dlyb_cr);
